Finish update the handler's code

+++ handler/security/models/service.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Service struct {
-	Id        int       `json:"id" gorm:"column:id"`
+	Id        int64     `json:"id" gorm:"column:id"`
 	Name      string    `json:"name" gorm:"column:name"`
 	CreatedAt time.Time `json:"created_at" gorm:"column:created_at"`
 	UpdatedAt time.Time `json:"updated_at" gorm:"column:updated_at"`
